product admin

+++ src/main/java/com/msstore/AdminController/ProductAdminController.java
@@ -6,12 +6,15 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,6 +28,8 @@ import com.msstore.DAO.SanPhamDAO;
 import com.msstore.Entity.HinhAnh;
 import com.msstore.Entity.LoaiSP;
 import com.msstore.Entity.SanPham;
+import com.msstore.Entity.TaiKhoan;
+import com.msstore.Service.SessionService;
 
 @Controller
 @RequestMapping("/admin")
@@ -39,15 +44,20 @@ public class ProductAdminController {
 	@Autowired
 	LoaiSPDAO lspDAO;
 	
+	@Autowired
+	SessionService sessionService;
+	
 	@GetMapping("/product")
-	public String doGetProductAdmin(Model model,@RequestParam("page") Optional<Integer> page) {
-		int currentPage = page.orElse(0);
-		Pageable pageAble = PageRequest.of(currentPage, 5);
+	public String doGetProductAdmin(Model model,@RequestParam("page") Optional<Integer> page,
+			@ModelAttribute("product") SanPham product) {
+		Pageable pageAble = PageRequest.of(page.orElse(0), 5);
 		Page<SanPham> listProduct = spDAO.findAll(pageAble);
+		List<LoaiSP> categoris = lspDAO.findAll();
 		List<HinhAnh> img = haDAO.findAll();
 		Set<Long> nameSet = new HashSet<>();
 		img = img.stream().filter(e -> nameSet.add(e.getSp().getMaSP())).collect(Collectors.toList());
 		
+		model.addAttribute("categoris", categoris);
 		model.addAttribute("products", listProduct);
 		model.addAttribute("images", img);
 		return "admin/product";
@@ -55,13 +65,14 @@ public class ProductAdminController {
 	
 	@GetMapping("/product/edit/{productId}")
 	public String doEditProduct(Model model,@RequestParam("page") Optional<Integer> page, @PathVariable("productId") long productId) {
-		int currentPage = page.orElse(0);
-		Pageable pageAble = PageRequest.of(currentPage, 5);
+		Pageable pageAble = PageRequest.of(page.orElse(0), 5);
 		Page<SanPham> listProduct = spDAO.findAll(pageAble);
+		
 		List<LoaiSP> categoris = lspDAO.findAll();
 		List<HinhAnh> img = haDAO.findAll();
 		Set<Long> nameSet = new HashSet<>();
 		img = img.stream().filter(e -> nameSet.add(e.getSp().getMaSP())).collect(Collectors.toList());
+		
 		model.addAttribute("images", img);
 		model.addAttribute("categoris", categoris);
 		model.addAttribute("products", listProduct);
@@ -72,23 +83,92 @@ public class ProductAdminController {
 	}
 	
 	@PostMapping("/product/create")
-	public String doPostCreate(@ModelAttribute("product") SanPham product) {
-		spDAO.save(product);
+	public String doPostCreate(@Valid @ModelAttribute("product") SanPham product,BindingResult bindingResult,
+			@RequestParam("page") Optional<Integer> page, Model model) {
+		Pageable pageAble = PageRequest.of(page.orElse(0), 5);
+		Page<SanPham> listProduct = spDAO.findAll(pageAble);
+		
+		List<LoaiSP> categoris = lspDAO.findAll();
+		List<HinhAnh> img = haDAO.findAll();
+		Set<Long> nameSet = new HashSet<>();
+		img = img.stream().filter(e -> nameSet.add(e.getSp().getMaSP())).collect(Collectors.toList());
+		
+		if(bindingResult.hasErrors()) {
+			model.addAttribute("images", img);
+			model.addAttribute("categoris", categoris);
+			model.addAttribute("products", listProduct);
+			return"admin/product";
+		}else {
+			spDAO.save(product);
+		}
 		return "redirect:/admin/product";
 	}
 	
 	@RequestMapping("/product/update/{productId}")
-	public String doPostUpdate(@ModelAttribute("product") SanPham product,@PathVariable("productId") long productId) {
-		product.setMaSP(productId);
-		System.out.println(product.getMaSP());
-		spDAO.save(product);
-		return "redirect:/admin/product/edit/" + product.getMaSP();
+	public String doPostUpdate(@Valid @ModelAttribute("product") SanPham product, BindingResult bindingResult,
+			@PathVariable("productId") long productId, @RequestParam("page") Optional<Integer> page, Model model) {
+		Pageable pageAble = PageRequest.of(page.orElse(0), 5);
+		Page<SanPham> listProduct = spDAO.findAll(pageAble);
+		
+		List<LoaiSP> categoris = lspDAO.findAll();
+		List<HinhAnh> img = haDAO.findAll();
+		Set<Long> nameSet = new HashSet<>();
+		img = img.stream().filter(e -> nameSet.add(e.getSp().getMaSP())).collect(Collectors.toList());
+		
+		if(!spDAO.existsById(productId)) {
+			model.addAttribute("images", img);
+			model.addAttribute("categoris", categoris);
+			model.addAttribute("products", listProduct);
+			model.addAttribute("message", "Sản phẩm không tồn tại");
+			return"admin/product";
+		}else if(bindingResult.hasErrors()){
+			model.addAttribute("images", img);
+			model.addAttribute("categoris", categoris);
+			model.addAttribute("products", listProduct);
+			return"admin/product";
+		}else {
+			product.setMaSP(productId);
+			spDAO.save(product);
+			return "redirect:/admin/product/edit/" + product.getMaSP();
+		}
 	}
 	
+	@SuppressWarnings("unlikely-arg-type")
 	@RequestMapping("/product/delete/{productId}")
-	public String doPostDelete(@PathVariable("productId") long productId) {
-		spDAO.deleteById(productId);
-		return "redirect:/admin/product";
+	public String doPostDelete(@PathVariable("productId") Long productId,Model model,
+			@ModelAttribute("product") SanPham product,@RequestParam("page") Optional<Integer> page) {
+		Pageable pageAble = PageRequest.of(page.orElse(0), 5);
+		Page<SanPham> listProduct = spDAO.findAll(pageAble);
+		
+		List<LoaiSP> categoris = lspDAO.findAll();
+		List<HinhAnh> img = haDAO.findAll();
+		Set<Long> nameSet = new HashSet<>();
+		img = img.stream().filter(e -> nameSet.add(e.getSp().getMaSP())).collect(Collectors.toList());
+		
+		if(productId.equals(0)) {
+			model.addAttribute("images", img);
+			model.addAttribute("categoris", categoris);
+			model.addAttribute("products", listProduct);
+			model.addAttribute("message", "Vui lòng chọn sản phẩm để xóa");
+			return"admin/product";
+		}else {
+			spDAO.deleteById(productId);
+			return "redirect:/admin/product";
+		}
+
+	}
+	
+	@RequestMapping("/product/search")
+	public String Search(Model model,@RequestParam("page") Optional<Integer> page,
+			@ModelAttribute("product") SanPham product, @RequestParam("Keyword") Optional<String> kw) {
+		String kwords = kw.orElse(sessionService.get("keywords",""));
+		sessionService.get("keywords",kwords);
+		Pageable pageable = PageRequest.of(page.orElse(0), 10);
+		
+		Page<SanPham> sp = spDAO.findAllByKeywords(pageable, "%" + kwords + "%");
+		model.addAttribute("products", sp);
+
+		return"admin/product";		
 	}
 	
 	@RequestMapping("/product/reset")
